Avoid circular import: move cleanup method to training.py
