+++ sockeye/arguments.py
@@ -24,7 +24,6 @@ import yaml
 
 from . import constants as C
 from . import data_io
-from . import utils
 
 
 class ConfigArgumentParser(argparse.ArgumentParser):
@@ -1085,26 +1084,28 @@ def add_training_args(params):
                               action="store_true",
                               help='In addition to keeping the last n params files, also keep params from checkpoint 0.')
 
-    
     train_params.add_argument('--cache-last-best-params',
                               required=False,
                               type=int,
                               default=0,
-                              help='Cache the last n best params files, as distinct from the last n in sequence. Use 0 or negative to disable. Default: %(default)s')
+                              help='Cache the last n best params files, as distinct from the last n in sequence. '
+                                   'Use 0 or negative to disable. Default: %(default)s')
 
     train_params.add_argument('--cache-strategy',
                               required=False,
                               type=str,
                               default=C.AVERAGE_BEST,
                               choices=C.AVERAGE_CHOICES,
-                              help='Strategy to use when deciding which are the "best" params files. Default: %(default)s')
+                              help='Strategy to use when deciding which are the "best" params files. '
+                                   'Default: %(default)s')
 
     train_params.add_argument('--cache-metric',
                               required=False,
                               type=str,
                               default=C.PERPLEXITY,
                               choices=C.METRICS,
-                              help='Metric to use when deciding which are the "best" params files. Default: %(default)s')
+                              help='Metric to use when deciding which are the "best" params files. '
+                                   'Default: %(default)s')
 
     train_params.add_argument('--dry-run',
                               action='store_true',
